Correct many lint issues, partition the code for manageability
