+++ onto_tool/mdutils.py
@@ -5,14 +5,13 @@ In the first instance this is to HTML5
 The class md2html() has a template that calls bootstrap for styling
 """
 
-import jinja2
-import markdown2
 import io
 
-class Markdown2HTML:
+import jinja2
+import markdown2
 
-    def __init__(self):
-        self.TEMPLATE = """<!DOCTYPE html>
+
+_TEMPLATE = """<!DOCTYPE html>
 <html>
 <head>
     <link href="http://netdna.bootstrapcdn.com/twitter-bootstrap/2.3.0/css/bootstrap-combined.min.css" rel="stylesheet">
@@ -41,15 +40,16 @@ class Markdown2HTML:
 </html>
 """
 
-    def md2html(self, md):
-        """
-        Parameters:
-        md (str): the markdown text that is to be converted to HTML5
-        Returns:
-        docfile (file object): a file object containing the HTML5
-        """
-        extensions = ['extra', 'smarty', 'tables']
-        html = markdown2.markdown(md, extras=extensions)
-        doc = jinja2.Template(self.TEMPLATE).render(content=html)
-        docfile = io.StringIO(doc)
-        return docfile
+
+def md2html(md):
+    """
+    Parameters:
+    md (str): the markdown text that is to be converted to HTML5
+    Returns:
+    docfile (file object): a file object containing the HTML5
+    """
+    extensions = ['extra', 'smarty', 'tables']
+    html = markdown2.markdown(md, extras=extensions)
+    doc = jinja2.Template(_TEMPLATE).render(content=html)
+    docfile = io.StringIO(doc)
+    return docfile
