+++ onto_tool/ontograph.py
@@ -14,23 +14,25 @@ import os
 import re
 import sys
 from collections import defaultdict
-from math import log, pow
+from math import log
 from string import Template
 from time import perf_counter
 from urllib.parse import urlparse, urlunparse
 
 import pydot
-from rdflib import Graph, BNode
-from rdflib.namespace import RDF, OWL
+from rdflib import BNode, Graph
+from rdflib.namespace import OWL, RDF
 from rdflib.util import guess_format
 
 from .sparql_utils import create_endpoint, select_query
 
-
 # Ignore \l - uses them as a line separator
 # pylint: disable=W1401
 
+
 class OntoGraf:
+    """Generates schema or instance graphs."""
+
     def __init__(self, files, repo=None, **kwargs):
         self.wee = kwargs.get('wee', None)
         self.hide = kwargs.get('hide')
@@ -39,26 +41,8 @@ class OntoGraf:
         if not version:
             version = datetime.datetime.now().isoformat()[:10]
 
-        self.cached_data = {}
-        if kwargs.get('cache'):
-            self.cache = json.load(kwargs['cache'])
-            self.save_cache = None # duplicative
-        else:
-            self.cache = None
-            self.save_cache = kwargs.get('save_cache')
-
-        if self.cache:
-            self.title = self.cache['title']
-            out_filename = self.cache['out_filename']
-        elif repo:
-            anonymized, repo_base = self.anonymize_url(repo)
-            self.title = (title or anonymized) + ': ' + version
-            out_filename = repo_base
-            self.cached_data['title'] = 'Cached ' + self.title
-            self.cached_data['out_filename'] = 'cached_' + out_filename
-        else:
-            self.title = f'{title or "Gist"} Ontology: {version}'
-            out_filename = f'{title or "Gist"}{version}'
+        out_filename = self.__configure_data_source(
+            repo, kwargs, title, version)
         self.graf = None
         self.files = files
         self.repo = repo
@@ -78,10 +62,12 @@ class OntoGraf:
 
         self.limit = kwargs.get('limit', 500000)
         self.threshold = kwargs.get('threshold', 10)
-        self.node_data = {}
-        self.class_names = {}
-        self.class_counts = defaultdict(int)
-        self.inheritance = []
+
+        # Data structures for instance data gather
+        self.node_data, self.class_names, self.class_counts, self.inheritance = \
+            {}, {}, defaultdict(int), []
+
+        # Default color configuration
         self.super_color = "blue"
         self.arrow_color = "darkorange2"
         self.shacl_color = "darkgreen"
@@ -99,8 +85,32 @@ class OntoGraf:
         self.show_shacl = kwargs.get('show_shacl')
         self.shapes = defaultdict(list)
 
+    def __configure_data_source(self, repo, kwargs, title, version):
+        """Determine graph title and output location from data source."""
+        self.cached_data = {}
+        if kwargs.get('cache'):
+            self.cache = json.load(kwargs['cache'])
+            self.save_cache = None  # duplicative
+        else:
+            self.cache = None
+            self.save_cache = kwargs.get('save_cache')
+
+        if self.cache:
+            self.title = self.cache['title']
+            out_filename = self.cache['out_filename']
+        elif repo:
+            anonymized, repo_base = self.__anonymize_url(repo)
+            self.title = (title or anonymized) + ': ' + version
+            out_filename = repo_base
+            self.cached_data['title'] = 'Cached ' + self.title
+            self.cached_data['out_filename'] = 'cached_' + out_filename
+        else:
+            self.title = f'{title or "Gist"} Ontology: {version}'
+            out_filename = f'{title or "Gist"}{version}'
+        return out_filename
+
     @staticmethod
-    def anonymize_url(url):
+    def __anonymize_url(url):
         """Remove username and password from URI, if present."""
         parsed = urlparse(url)
         anonymized = urlunparse((parsed.scheme,
@@ -109,7 +119,7 @@ class OntoGraf:
                                  '', '', ''))
         return anonymized, os.path.basename(parsed.path)
 
-    def remote_select_query(self, query):
+    def __remote_select_query(self, query):
         """Execute SPARQL SELECT query, return results as generator."""
         logging.debug(f"Query against {self.repo}")
         logging.debug(f"Query\n {query}")
@@ -117,7 +127,7 @@ class OntoGraf:
         sparql = create_endpoint(self.repo)
         return select_query(sparql, query)
 
-    def local_select_query(self, query):
+    def __local_select_query(self, query):
         """Execute SPARQL SELECT query on local data, return results as generator."""
         logging.debug(f"Local Query\n {query}")
         if not self.data:
@@ -129,20 +139,20 @@ class OntoGraf:
 
         results = self.data.query(query)
         for result in results:
-            yield dict((str(k), str(v) if v is not None else None) for k, v in zip(results.vars, result))
+            yield dict((str(k), str(v) if v is not None else None)
+                       for k, v in zip(results.vars, result))
 
-    def select_query(self, query:str, id: str):
+    def __select_query(self, query: str, cache_id: str):
         if self.cache:
-            return self.cache[id]
+            return self.cache[cache_id]
         if self.repo:
-            data = list(self.remote_select_query(query))
-            self.cached_data[id] = data
+            data = list(self.__remote_select_query(query))
+            self.cached_data[cache_id] = data
             return data
-        else:
-            return list(self.local_select_query(query))
+        return list(self.__local_select_query(query))
 
     @staticmethod
-    def strip_uri(uri):
+    def __strip_uri(uri):
         stripped = re.sub(r'^.*[/#](.*?)(X.x.x|\d+.\d+.\d+)?$', '\\1',
                           re.sub(r'[#/]$', '', str(uri)))
         if not stripped:
@@ -151,6 +161,7 @@ class OntoGraf:
         return stripped
 
     def gather_schema_info_from_files(self):
+        """Load schema data from local ontology."""
         self.node_data = {}
         for file_path in self.files:
             filename = os.path.basename(file_path)
@@ -159,19 +170,22 @@ class OntoGraf:
             graph.parse(file_path, format=guess_format(file_path))
 
             ontology = next(graph.subjects(RDF.type, OWL.Ontology))
-            ontology_name = self.strip_uri(ontology)
-            classes = [self.strip_uri(c) for c in graph.subjects(RDF.type, OWL.Class)
+            ontology_name = self.__strip_uri(ontology)
+            classes = [self.__strip_uri(c) for c in graph.subjects(RDF.type, OWL.Class)
                        if not isinstance(c, BNode)]
-            obj_props = [self.strip_uri(c)
+            obj_props = [self.__strip_uri(c)
                          for c in graph.subjects(RDF.type, OWL.ObjectProperty)]
-            data_props = [self.strip_uri(c)
+            data_props = [self.__strip_uri(c)
                           for c in graph.subjects(RDF.type, OWL.DatatypeProperty)]
-            annotation_props = [self.strip_uri(c)
+            annotation_props = [self.__strip_uri(c)
                                 for c in graph.subjects(RDF.type, OWL.AnnotationProperty)]
             all_seen = set(classes + obj_props + data_props + annotation_props)
-            gist_things = [self.strip_uri(s) for (s, o) in graph.subject_objects(RDF.type)
-                           if not isinstance(s, BNode) and not s == ontology and not self.strip_uri(s) in all_seen]
-            imports = [self.strip_uri(c) for c in graph.objects(ontology, OWL.imports)]
+            gist_things = [
+                self.__strip_uri(s) for (s, o) in graph.subject_objects(RDF.type)
+                if not isinstance(s, BNode) and not s == ontology and not self.__strip_uri(s)
+                in all_seen]
+            imports = [self.__strip_uri(c)
+                       for c in graph.objects(ontology, OWL.imports)]
 
             self.node_data[filename] = {
                 "ontology": ontology,
@@ -186,6 +200,7 @@ class OntoGraf:
         return self.node_data
 
     def gather_schema_info_from_repo(self):
+        """Load schema data from SPARQL endpoint."""
         onto_data = defaultdict(lambda: defaultdict(list))
         if self.single_graph:
             onto_query = """
@@ -239,26 +254,31 @@ class OntoGraf:
             str(OWL.AnnotationProperty): 'annotation_propertiesList',
             str(OWL.imports): 'imports'
         }
-        for entity in self.remote_select_query(onto_query):
+        for entity in self.__remote_select_query(onto_query):
             key = mapping.get(entity['type'], 'gist_thingsList')
-            onto_data[entity['ontology']][key].append(self.strip_uri(entity['entity']))
+            onto_data[entity['ontology']][key].append(
+                self.__strip_uri(entity['entity']))
 
         if not onto_data:
-            logging.warning('Could not find any ontology entities in %s', self.repo)
+            logging.warning(
+                'Could not find any ontology entities in %s', self.repo)
             return
 
         self.node_data = defaultdict(dict)
         for ontology, props in onto_data.items():
             self.node_data[ontology]['ontology'] = ontology
-            self.node_data[ontology]['ontologyName'] = self.strip_uri(ontology)
+            self.node_data[ontology]['ontologyName'] = self.__strip_uri(
+                ontology)
             for key in set(mapping.values()).union({'gist_thingsList'}):
                 if key != 'imports':
-                    self.node_data[ontology][key] = "\\l".join(sorted(set(props[key]))) if key in props else ''
+                    self.node_data[ontology][key] = "\\l".join(
+                        sorted(set(props[key]))) if key in props else ''
                 else:
-                    self.node_data[ontology][key] = props[key] if key in props else []
-        return self.node_data
+                    self.node_data[ontology][key] = props[key] if key in props else [
+                    ]
 
     def create_schema_graf(self):
+        """Create schema graph from local or remote data."""
         data_dict = self.node_data
         wee = self.wee
         # When 'wee' is not specified at all, it will be None
@@ -287,18 +307,12 @@ class OntoGraf:
         for file, file_data in data_dict.items():
             if file != '':
                 ontology = file_data["ontology"]
-                if not self.ontology_matches_filter(ontology):
+                if not self.__ontology_matches_filter(ontology):
                     logging.debug("Filtered out %s", ontology)
                     continue
                 ontology_name = file_data["ontologyName"]
-                classes = file_data["classesList"]
-                obj_properties = file_data["obj_propertiesList"]
-                data_properties = file_data["data_propertiesList"]
-                annotation_properties = file_data["annotation_propertiesList"]
-                gist_things = file_data["gist_thingsList"]
-                imports = file_data["imports"]
                 render_compact = wee is not None and (
-                        not wee or any(re.search(pat, ontology) for pat in wee)
+                    not wee or any(re.search(pat, ontology) for pat in wee)
                 )
                 if render_compact:
                     node = pydot.Node(ontology_name)
@@ -306,67 +320,84 @@ class OntoGraf:
                     ontology_info = "{{{}\\l\\l{}|{}|{}|{}|{}|{}}}".format(
                         file,
                         ontology_name,
-                        classes,
-                        obj_properties,
-                        data_properties,
-                        annotation_properties,
-                        gist_things)
+                        file_data["classesList"],
+                        file_data["obj_propertiesList"],
+                        file_data["data_propertiesList"],
+                        file_data["annotation_propertiesList"],
+                        file_data["gist_thingsList"])
                     node = pydot.Node(ontology_name,
                                       label=ontology_info)
 
                 self.graf.add_node(node)
 
-                for imported in imports:
+                for imported in file_data["imports"]:
                     edge = pydot.Edge(ontology_name, imported,
                                       color=self.arrow_color,
                                       arrowhead=self.arrowhead)
                     self.graf.add_edge(edge)
         self.graf.write(self.outdot)
         if not self.no_image:
-            self.graf.write_png(self.outpng)
+            self.graf.write(self.outpng, format='png')
         logging.debug("Plots saved")
 
-    # Print iterations progress
-    @staticmethod
-    def print_progress_bar(iteration, total,
-                           prefix='', suffix='', decimals=1, length=100,
-                           fill='X', print_end="\r"):
-        """
+    class ProgressBar():
+        """Renders a progress bar in a terminal.
+
         Stolen from https://stackoverflow.com/questions/3173320/text-progress-bar-in-the-console
-        Call in a loop to create terminal progress bar
-        @params:
-            iteration   - Required  : current iteration (Int)
-            total       - Required  : total iterations (Int)
-            prefix      - Optional  : prefix string (Str)
-            suffix      - Optional  : suffix string (Str)
-            decimals    - Optional  : positive number of decimals in percent complete (Int)
-            length      - Optional  : character length of bar (Int)
-            fill        - Optional  : bar fill character (Str)
-            print_end   - Optional  : end character (e.g. "\r", "\r\n") (Str)
         """
-        # if not sys.stdout.isatty():
-        #     return
-        if logging.root.getEffectiveLevel() == logging.DEBUG:
-            return
 
-        percent = ("{0:." + str(decimals) + "f}").format(100 * (iteration / float(total)))
-        filled_length = int(length * iteration // total)
-        bar = fill * filled_length + '-' * (length - filled_length)
-        print(f'\r{prefix} |{bar}| {percent}% {suffix}', end=print_end)
-        # Print New Line on Complete
-        if iteration == total:
-            print()
-        sys.stdout.flush()
+        def __init__(self, total, **kwargs):
+            """
+            @params:
+                total       - Required  : total iterations (Int)
+                prefix      - Optional  : prefix string (Str, default='')
+                suffix      - Optional  : suffix string (Str, default='')
+                decimals    - Optional  : positive number of decimals
+                                          in percent complete (Int, default=1)
+                length      - Optional  : character length of bar (Int, default=100)
+                fill        - Optional  : bar fill character (Str, default='X')
+                print_end   - Optional  : end character (e.g. "\r", "\r\n") (Str, default='\r')
+            """
+            self.total = total
+            self.prefix = kwargs.get('prefix', '')
+            self.suffix = kwargs.get('suffix', '')
+            self.decimals = kwargs.get('decimals', 1)
+            self.length = kwargs.get('length', 100)
+            self.fill = kwargs.get('fill', 'X')
+            self.print_end = kwargs.get('print_end', '\r')
 
-    def hidden(self, uri):
+        def print(self, iteration, suffix=None):
+            """Call in a loop to create terminal progress bar."""
+            # if not sys.stdout.isatty():
+            #     return
+            if logging.root.getEffectiveLevel() == logging.DEBUG:
+                return
+
+            if not suffix:
+                suffix = self.suffix
+            percent = ("{0:." + str(self.decimals) + "f}").format(100 *
+                                                                  (iteration / float(self.total)))
+            filled_length = int(self.length * iteration // self.total)
+            progress_bar = self.fill * filled_length + \
+                '-' * (self.length - filled_length)
+            print(
+                f'\r{self.prefix} |{progress_bar}| {percent}% {suffix}', end=self.print_end)
+            # Print New Line on Complete
+            if iteration == self.total:
+                print()
+            sys.stdout.flush()
+
+    def __hidden(self, uri):
         return any(re.search(pat, uri) for pat in self.hide)
 
-    def deepest_class(self, class_uris: str):
-        _, deepest = max((self.inheritance.index(cls_uri) if cls_uri in self.inheritance else -1, cls_uri)
-                         for cls_uri in class_uris.split(' '))
+    def __deepest_class(self, class_uris: str):
+        _, deepest = max((
+            self.inheritance.index(cls_uri) if cls_uri in self.inheritance else -1, cls_uri)
+            for cls_uri in class_uris.split(' '))
         return deepest
 
     def gather_instance_info(self):
+        """Gather class and predicate usage from local data or SPARQL endpoint."""
         predicate_query = Template("""
         prefix owl: <http://www.w3.org/2002/07/owl#>
         prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
@@ -379,7 +410,9 @@ class OntoGraf:
           {
             select distinct ?predicate where {
               ?s ?predicate ?o
-              FILTER(?predicate NOT IN (rdf:type, rdfs:label, skos:prefLabel, skos:altLabel, skos:definition))
+              FILTER(?predicate NOT IN (
+                  rdf:type, rdfs:label, skos:prefLabel,
+                  skos:altLabel, skos:definition))
             }
           }
           FILTER (!STRSTARTS(STR(?predicate), 'http://www.w3.org/2002/07/owl#'))
@@ -396,52 +429,58 @@ class OntoGraf:
         }
         """).substitute(language=self.label_lang)
         self.node_data = {}
-        all_predicates = self.select_query(predicate_query, 'all_predicates')
+        all_predicates = self.__select_query(predicate_query, 'all_predicates')
 
         hidden_predicates = set(predicate['predicate'] for predicate in all_predicates
-                                if self.hidden(predicate['predicate']))
+                                if self.__hidden(predicate['predicate']))
         logging.debug("Hiding predicates: %s", hidden_predicates)
-        all_predicates = [predicate for predicate in all_predicates if predicate['predicate'] not in hidden_predicates]
+        all_predicates = [
+            predicate for predicate in all_predicates
+            if predicate['predicate'] not in hidden_predicates]
 
         if not all_predicates:
-            logging.warning('No interesting predicates found in %s', self.repo or ' specified files')
+            logging.warning('No interesting predicates found in %s',
+                            self.repo or ' specified files')
             return
 
-        self.build_class_hierarchy()
+        self.__build_class_hierarchy()
 
+        progress_bar = self.ProgressBar(len(all_predicates),
+                                        prefix='Processing predicates:', suffix='', length=50)
         for count, predicate_row in enumerate(all_predicates):
             predicate = predicate_row['predicate']
             predicate_str = predicate_row['label'] if predicate_row.get('label') \
-                else self.strip_uri(predicate)
+                else self.__strip_uri(predicate)
 
-            self.print_progress_bar(count, len(all_predicates),
-                                    prefix='Processing predicates:',
-                                    suffix=predicate_str + ' ' * 20, length=50)
+            progress_bar.print(count,
+                               suffix=predicate_str + ' ' * 20)
             pre_time = perf_counter()
-            query_text = self.create_predicate_query(predicate, predicate_row.get('type'), self.limit)
-            predicate_usage = list(self.select_query(query_text, predicate))
-            logging.debug("%s items returned for %s", len(predicate_usage), predicate)
+            query_text = self.__create_predicate_query(
+                predicate, predicate_row.get('type'), self.limit)
+            predicate_usage = list(self.__select_query(query_text, predicate))
+            logging.debug("%s items returned for %s",
+                          len(predicate_usage), predicate)
             for usage in predicate_usage:
                 if 'tgt' in usage and usage['tgt'] is None:
                     del usage['tgt']
-                if 'src' not in usage or usage['src'] is None or int(usage.get('num', 0)) < self.threshold:
+                if usage.get('src') is None or int(usage.get('num', 0)) < self.threshold:
                     continue
-                self.record_predicate_usage(predicate, predicate_str, usage)
+                self.__record_predicate_usage(predicate, predicate_str, usage)
 
-            logging.debug("Fetching %s took %d seconds", str(predicate_row), perf_counter() - pre_time)
+            logging.debug("Fetching %s took %d seconds", str(
+                predicate_row), perf_counter() - pre_time)
 
-        self.print_progress_bar(len(all_predicates), len(all_predicates),
-                                prefix='Processing predicates:', suffix='Complete', length=50)
+        progress_bar.print(len(all_predicates), suffix='Complete')
 
-        self.prune_for_inheritance()
+        self.__prune_for_inheritance()
 
         if self.show_shacl:
-            self.add_shacl_coloring()
+            self.__add_shacl_coloring()
 
         if self.save_cache:
             json.dump(self.cached_data, self.save_cache)
 
-    def build_class_hierarchy(self):
+    def __build_class_hierarchy(self):
         inheritance_query = Template("""
         prefix owl: <http://www.w3.org/2002/07/owl#>
         prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
@@ -454,7 +493,9 @@ class OntoGraf:
           }
           UNION
           {
-             ?class (owl:equivalentClass|rdfs:subClassOf)/(owl:unionOf|owl:intersectionOf)/rdf:rest*/rdf:first ?parent .
+             ?class (owl:equivalentClass|rdfs:subClassOf)/
+                    (owl:unionOf|owl:intersectionOf)/
+                    rdf:rest*/rdf:first ?parent .
             ?parent a owl:Class
           }
           filter (!isblank(?class) && !isblank(?parent))
@@ -468,14 +509,17 @@ class OntoGraf:
           }
         }
         """).substitute(language=self.label_lang)
-        parents = self.select_query(inheritance_query, 'parents')
+        parents = self.__select_query(inheritance_query, 'parents')
 
         for inheritance_info in parents:
-            self.superclasses[inheritance_info['class']].add(inheritance_info['parent'])
+            self.superclasses[inheritance_info['class']].add(
+                inheritance_info['parent'])
             self.class_names[inheritance_info['class']] = \
-                inheritance_info.get('c_label') or self.strip_uri(inheritance_info['class'])
+                inheritance_info.get('c_label') or self.__strip_uri(
+                    inheritance_info['class'])
             self.class_names[inheritance_info['parent']] = \
-                inheritance_info.get('p_label') or self.strip_uri(inheritance_info['parent'])
+                inheritance_info.get('p_label') or self.__strip_uri(
+                    inheritance_info['parent'])
 
         # Determine evaluation order, root classes to leaves
         remaining_classes = set(self.superclasses.keys())
@@ -490,49 +534,56 @@ class OntoGraf:
             for cls in next_set:
                 parents = set(self.superclasses[cls])
                 for parent in parents:
-                    self.superclasses[cls].update(self.superclasses.get(parent, set()))
+                    self.superclasses[cls].update(
+                        self.superclasses.get(parent, set()))
             eval_order.extend(next_set)
             remaining_classes.difference_update(next_set)
 
-        logging.debug('Inheritance evaluation order:\n%s',
-                      "\n".join(
-                          f"\t{self.strip_uri(cls)}: {list(self.strip_uri(sup) for sup in self.superclasses[cls])}"
-                          for cls in eval_order))
+        logging.debug(
+            'Inheritance evaluation order:\n%s',
+            "\n".join(
+                f"\t{self.__strip_uri(cls)}: {list(self.__strip_uri(sup) for sup in self.superclasses[cls])}"
+                for cls in eval_order))
         self.inheritance = eval_order
 
-        class_query = self.create_class_count_query(self.limit)
-        class_counts = self.select_query(class_query, 'class_counts')
+        class_query = self.__create_class_count_query(self.limit)
+        class_counts = self.__select_query(class_query, 'class_counts')
 
         for instance_info in class_counts:
-            self.class_counts[self.deepest_class(instance_info['src'])] += int(instance_info['num'])
+            self.class_counts[self.__deepest_class(
+                instance_info['src'])] += int(instance_info['num'])
 
-    def prune_for_inheritance(self):
+    def __prune_for_inheritance(self):
         for cls in reversed(self.inheritance):
             if cls not in self.node_data or not self.superclasses[cls]:
                 continue
-            self.merge_with_parent(cls, 'links')
-            self.merge_with_parent(cls, 'data')
+            self.__merge_with_parent(cls, 'links')
+            self.__merge_with_parent(cls, 'data')
 
-    def merge_with_parent(self, cls, link_type):
+    def __merge_with_parent(self, cls, link_type):
         removed = []
         for link, count in self.node_data[cls][link_type].items():
             # Locate parents with same link
-            parents_with_link = list(parent for parent in self.superclasses[cls]
-                                     if parent in self.node_data and link in self.node_data[parent][link_type])
+            parents_with_link = list(
+                parent for parent in self.superclasses[cls]
+                if parent in self.node_data and link in self.node_data[parent][link_type]
+            )
             if not parents_with_link:
                 continue
             # Move to the most immediate parent, to create proper superclass chains
-            _, max_parent_with_link = max((self.inheritance.index(parent), parent) for parent in parents_with_link)
+            _, max_parent_with_link = max((self.inheritance.index(
+                parent), parent) for parent in parents_with_link)
             self.node_data[max_parent_with_link][link_type][link] += count
             self.node_data[cls]['supers'].add(max_parent_with_link)
             removed.append(link)
         for r in removed:
             del self.node_data[cls][link_type][r]
 
-    def record_predicate_usage(self, predicate, predicate_str, usage):
-        src_uri = self.deepest_class(usage['src'])
-        tgt_uri = self.deepest_class(usage['tgt']) if 'tgt' in usage else None
-        if self.hidden(src_uri) or (tgt_uri is not None and self.hidden(tgt_uri)):
+    def __record_predicate_usage(self, predicate, predicate_str, usage):
+        src_uri = self.__deepest_class(usage['src'])
+        tgt_uri = self.__deepest_class(
+            usage['tgt']) if 'tgt' in usage else None
+        if self.__hidden(src_uri) or (tgt_uri is not None and self.__hidden(tgt_uri)):
             if tgt_uri is not None:
                 logging.debug('Hiding %s to %s link', src_uri, tgt_uri)
             else:
@@ -540,9 +591,9 @@ class OntoGraf:
             return
         if src_uri not in self.node_data:
             if src_uri is None:
-                raise Exception("None src_uri in " + str(usage))
+                raise LookupError("None src_uri in " + str(usage))
             src = {
-                'label': self.class_names.get(src_uri, self.strip_uri(src_uri)),
+                'label': self.class_names.get(src_uri, self.__strip_uri(src_uri)),
                 'links': defaultdict(int),
                 'data': defaultdict(int),
                 'supers': set()
@@ -552,21 +603,22 @@ class OntoGraf:
             src = self.node_data[src_uri]
         if usage.get('dt'):
             src['data'][(predicate,
-                         predicate_str or self.strip_uri(predicate),
-                         self.strip_uri(usage['dt']))] += int(usage['num'])
+                         predicate_str or self.__strip_uri(predicate),
+                         self.__strip_uri(usage['dt']))] += int(usage['num'])
         else:
             if tgt_uri not in self.node_data:
                 if tgt_uri is None:
-                    raise Exception("None tgt_uri in " + str(usage))
+                    raise LookupError("None tgt_uri in " + str(usage))
                 self.node_data[tgt_uri] = {
-                    'label': self.class_names.get(tgt_uri, self.strip_uri(tgt_uri)),
+                    'label': self.class_names.get(tgt_uri, self.__strip_uri(tgt_uri)),
                     'links': defaultdict(int),
                     'data': defaultdict(int),
                     'supers': set()
                 }
-            src['links'][(predicate, predicate_str, tgt_uri)] += int(usage['num'])
+            src['links'][(predicate, predicate_str, tgt_uri)
+                         ] += int(usage['num'])
 
-    def add_shacl_coloring(self):
+    def __add_shacl_coloring(self):
         shacl_query = """
             prefix sh: <http://www.w3.org/ns/shacl#>
 
@@ -577,11 +629,11 @@ class OntoGraf:
               { ?shape (sh:and|sh:or|sh:xone|sh:not|rdf:first|rdf:rest)+/sh:path ?property }
             }
             """
-        shacl_data = self.select_query(shacl_query, 'shacl')
+        shacl_data = self.__select_query(shacl_query, 'shacl')
         for row in shacl_data:
             self.shapes[row['class']].append(row['property'])
 
-    def create_class_count_query(self, limit):
+    def __create_class_count_query(self, limit):
         class_query = """
             prefix owl: <http://www.w3.org/2002/07/owl#>
             prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
@@ -601,11 +653,11 @@ class OntoGraf:
             } group by ?src
             """
         query_text = Template(class_query).substitute(
-            pattern=self.filtered_graph_pattern(str(RDF.type)),
+            pattern=self.__filtered_graph_pattern(str(RDF.type)),
             limit=limit)
         return query_text
 
-    def create_predicate_query(self, predicate, predicate_type, limit):
+    def __create_predicate_query(self, predicate, predicate_type, limit):
         if predicate_type == str(OWL.ObjectProperty):
             type_query = """
                 prefix owl: <http://www.w3.org/2002/07/owl#>
@@ -617,7 +669,10 @@ class OntoGraf:
 
                 select ?src ?tgt (COUNT(?src) as ?num) where {
                   {
-                    select (group_concat(?src_c;separator=' ') as ?src) (group_concat(?tgt_c;separator=' ') as ?tgt) where {
+                    select
+                        (group_concat(?src_c;separator=' ') as ?src)
+                        (group_concat(?tgt_c;separator=' ') as ?tgt)
+                    where {
                       $pattern
                       FILTER(!ISBLANK(?s))
                       ?s a ?src_c .
@@ -698,7 +753,7 @@ class OntoGraf:
                 }
                 """
         query_text = Template(type_query).substitute(
-            pattern=self.filtered_graph_pattern(predicate),
+            pattern=self.__filtered_graph_pattern(predicate),
             limit=limit)
         return query_text
 
@@ -706,7 +761,7 @@ class OntoGraf:
     MAX_LINE_WIDTH = 5
 
     @staticmethod
-    def line_width(num_used, graph_max):
+    def __line_width(num_used, graph_max):
         """Scale line width relative to the most commonly occurring edge for the graph"""
         if graph_max == 1:
             return OntoGraf.MAX_LINE_WIDTH
@@ -717,25 +772,29 @@ class OntoGraf:
                        max(OntoGraf.MIN_LINE_WIDTH,
                            round(log(num_used, pow(graph_max, 1.0 / OntoGraf.MAX_LINE_WIDTH)))))
         except ValueError:
-            logging.warning('Failed to determine line width from num=%d max=%d', num_used, graph_max)
+            logging.warning(
+                'Failed to determine line width from num=%d max=%d', num_used, graph_max)
             return OntoGraf.MIN_LINE_WIDTH
 
     MAX_FONT_SIZE = 24
     MIN_FONT_SIZE = 14
 
     @staticmethod
-    def font_size(num_instances, graph_max):
+    def __font_size(num_instances, graph_max):
         """Scale line width relative to the most commonly occurring edge for the graph"""
         if num_instances == 0 or graph_max == 1:
             return OntoGraf.MIN_FONT_SIZE
         span = OntoGraf.MAX_FONT_SIZE - OntoGraf.MIN_FONT_SIZE
         try:
-            return OntoGraf.MIN_FONT_SIZE - 1 + min(span, max(1, round(log(num_instances, pow(graph_max, 1.0 / span)))))
+            return OntoGraf.MIN_FONT_SIZE - 1 + \
+                min(span, max(1, round(log(num_instances, pow(graph_max, 1.0 / span)))))
         except ValueError:
-            logging.warning('Failed to determine font size from num=%d max=%d', num_instances, graph_max)
+            logging.warning(
+                'Failed to determine font size from num=%d max=%d', num_instances, graph_max)
             return OntoGraf.MIN_FONT_SIZE
 
     def create_instance_graf(self, data_dict=None):
+        """Create graph from gathered instance data."""
         self.graf = pydot.Dot(graph_type='digraph',
                               label=self.title,
                               labelloc='t',
@@ -758,7 +817,8 @@ class OntoGraf:
             logging.warning("No data found with the specified parameters.")
             return
 
-        # Determine the maximum number any edge occurs in the data, so the edge widths can be properly scaled
+        # Determine the maximum number any edge occurs in the data,
+        # so the edge widths can be properly scaled
         common_list = [
             occurs for class_data in data_dict.values()
             for occurs in class_data['links'].values()
@@ -768,11 +828,13 @@ class OntoGraf:
         max_instance = max(self.class_counts.values())
 
         for class_, class_data in data_dict.items():
-            node = self.create_instance_graph_node(max_instance, class_, class_data)
+            node = self.__create_instance_graph_node(
+                max_instance, class_, class_data)
 
             self.graf.add_node(node)
 
-            by_predicate, compacted_links = self.determine_compacted_links(class_, class_data)
+            by_predicate, compacted_links = self.__determine_compacted_links(
+                class_, class_data)
 
             for link, num in class_data['links'].items():
                 predicate, predicate_str, target = link
@@ -780,30 +842,15 @@ class OntoGraf:
                     continue
                 edge = pydot.Edge(class_, target,
                                   label=predicate_str,
-                                  penwidth=self.line_width(num, max_common),
-                                  color=self.shacl_color if predicate in self.shapes[class_] else self.arrow_color,
+                                  penwidth=self.__line_width(num, max_common),
+                                  color=self.shacl_color if predicate in self.shapes[
+                                      class_] else self.arrow_color,
                                   arrowhead=self.arrowhead)
                 self.graf.add_edge(edge)
 
             for predicate in compacted_links:
-                links = by_predicate[predicate]
-                shared_node_id = class_ + '_' + predicate
-                shared_node = pydot.Node(name='"' + shared_node_id + '"', shape='point', color="black")
-                self.graf.add_node(shared_node)
-                total_count = sum(class_data['links'][link] for link in links)
-                edge_color = self.shacl_color if predicate in self.shapes[class_] else self.arrow_color
-                predicate_label = next(link for link in links)[1]
-                self.graf.add_edge(
-                    pydot.Edge(class_, shared_node_id,
-                               label=predicate_label,
-                               penwidth=self.line_width(total_count, max_common), color=edge_color))
-                for link in links:
-                    _, _, target = link
-                    edge = pydot.Edge(shared_node_id, target,
-                                      penwidth=self.line_width(class_data['links'][link], max_common),
-                                      color=edge_color,
-                                      arrowhead=self.arrowhead)
-                    self.graf.add_edge(edge)
+                self.add_compacted_edges(
+                    max_common, class_, predicate, class_data, by_predicate[predicate])
 
             for super_class in class_data['supers']:
                 edge = pydot.Edge(class_, super_class,
@@ -814,10 +861,40 @@ class OntoGraf:
 
         self.graf.write(self.outdot, encoding="utf-8")
         if not self.no_image:
-            self.graf.write_png(self.outpng)
+            self.graf.write(self.outpng, format='png')
         logging.debug("Plots saved")
 
-    def determine_compacted_links(self, class_, class_data):
+    def add_compacted_edges(self, max_common: int, source_class: str, predicate: str,
+                            class_data: dict, links: dict):
+        """Add graph edges for compacted links.
+
+        For a class that is linked to many classes via the same predicate,
+        generate a single labeled edge to a 'point' intermediate node, then
+        add unlabeled edges from the itermediate node to the target nodes.
+        This makes the diagram less crowded and clearer.
+        """
+        shared_node_id = source_class + '_' + predicate
+        shared_node = pydot.Node(
+            name='"' + shared_node_id + '"', shape='point', color="black")
+        self.graf.add_node(shared_node)
+        total_count = sum(class_data['links'][link] for link in links)
+        edge_color = self.shacl_color if predicate in self.shapes[
+            source_class] else self.arrow_color
+        predicate_label = next(link for link in links)[1]
+        self.graf.add_edge(
+            pydot.Edge(source_class, shared_node_id,
+                       label=predicate_label,
+                       penwidth=self.__line_width(total_count, max_common), color=edge_color))
+        for link in links:
+            _, _, target = link
+            edge = pydot.Edge(shared_node_id, target,
+                              penwidth=self.__line_width(
+                                  class_data['links'][link], max_common),
+                              color=edge_color,
+                              arrowhead=self.arrowhead)
+            self.graf.add_edge(edge)
+
+    def __determine_compacted_links(self, class_, class_data):
         by_predicate = defaultdict(set)
         if self.concentrate_links != 0:
             for link in class_data['links']:
@@ -831,15 +908,22 @@ class OntoGraf:
 
         return by_predicate, compacted_links
 
-    def create_instance_graph_node(self, max_instance, class_, class_data):
-        node_font_size = self.font_size(self.class_counts[class_], max_instance)
-        node_line_width = self.line_width(self.class_counts[class_], max_instance)
+    def __create_instance_graph_node(self, max_instance, class_, class_data):
+        # Using .format is more readable here
+        # pylint: disable=C0209
+
+        node_font_size = self.__font_size(
+            self.class_counts[class_], max_instance)
+        node_line_width = self.__line_width(
+            self.class_counts[class_], max_instance)
+        class_label = class_data['label'] if class_data['label'] else self.__strip_uri(class_)
+
         if class_data['data']:
             formatted_label = '<font point-size="{fontsize}" color="{label_fg}">{class_label}</font>'.format(
-                        label_fg="white" if class_ in self.shapes else "black",
-                        fontsize=node_font_size,
-                        class_label=class_data['label'] if class_data['label'] else self.strip_uri(class_)
-                )
+                label_fg="white" if class_ in self.shapes else "black",
+                fontsize=node_font_size,
+                class_label=class_label
+            )
             class_info = \
                 """<<table color="black" border="{line_width}" cellspacing="0" cellborder="1">
                     <tr>
@@ -861,7 +945,7 @@ class OntoGraf:
                               label=class_info)
         else:
             node = pydot.Node(name='"' + class_ + '"',
-                              label=class_data['label'] if class_data['label'] else self.strip_uri(class_),
+                              label=class_label,
                               style='filled',
                               fontsize=node_font_size,
                               penwidth=node_line_width,
@@ -871,49 +955,47 @@ class OntoGraf:
 
         return node
 
-    def ontology_matches_filter(self, ontology):
+    def __ontology_matches_filter(self, ontology):
         if self.include:
             return ontology in self.include
-        elif self.exclude:
+        if self.exclude:
             return ontology not in self.exclude
-        elif self.include_pattern:
+        if self.include_pattern:
             return any(re.search(pattern, ontology) for pattern in self.include_pattern)
-        elif self.exclude_pattern:
+        if self.exclude_pattern:
             return not any(re.search(pattern, ontology) for pattern in self.exclude_pattern)
-        else:
-            return True
+        return True
 
-    def filtered_graph_pattern(self, predicate):
+    def __filtered_graph_pattern(self, predicate):
         if not self.repo:
             # Local files always go in the default graph
             return f'?s <{predicate}> ?o .'
-        elif self.include:
+        if self.include:
             return f"""
             VALUES ?graph {{{" ".join(f"<{i}>" for i in self.include)}}}
             GRAPH ?graph {{
                 ?s <{predicate}> ?o .
             }}
             """
-        elif self.exclude:
+        if self.exclude:
             return f"""
             GRAPH ?graph {{
                 ?s <{predicate}> ?o .
             }}
             FILTER (?graph NOT IN ({", ".join(f"<{e}>" for e in self.exclude)}))
             """
-        elif self.include_pattern:
+        if self.include_pattern:
             return f"""
             GRAPH ?graph {{
                 ?s <{predicate}> ?o .
             }}
             FILTER ({" || ".join(f"REGEX(STR(?graph), '{i}')" for i in self.include_pattern)})
             """
-        elif self.exclude_pattern:
+        if self.exclude_pattern:
             return f"""
             GRAPH ?graph {{
                 ?s <{predicate}> ?o .
             }}
             FILTER ({" && ".join(f"!REGEX(STR(?graph), '{e}')" for e in self.exclude_pattern)})
             """
-        else:
-            return f'?s <{predicate}> ?o .'
+        return f'?s <{predicate}> ?o .'
